% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-euclid_norm.R
\name{euclid_norm}
\alias{euclid_norm}
\title{Compute the euclidean distance}
\usage{
euclid_norm(x)
}
\arguments{
\item{x}{A vector of difference in x and y dimension}
}
\value{
Euclidean distance between points
}
\description{
Compute the euclidean distance
}
