% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_2d_instance.R
\name{generate_2d_instance}
\alias{generate_2d_instance}
\title{Generate a problem instance}
\usage{
generate_2d_instance(
  seed = 1,
  no_of_points = 50,
  interval = c(min = -10, max = 10),
  arv = c(min = 1/120, max = 1),
  ar_dist_type = "uniform"
)
}
\arguments{
\item{seed}{Seed to ensure reproducibility}

\item{no_of_points}{The number of demand points}

\item{interval}{A named vector of the range for the x and y axis}

\item{arv}{A named vector of the range in the arrival rate (measured in requests per minute)}

\item{ar_dist_type}{Whether the distribution of arrival rates follows a "uniform" or "reciprocal" distribution}
}
\value{
A list
}
\description{
Generate a problem instance
}
\examples{
generate_2d_instance()
}
