% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-experiment.R
\name{generate_instances}
\alias{generate_instances}
\title{Generate instances for experiment}
\usage{
generate_instances(
  n = 5,
  no_of_points = 100,
  ar_mean = 0.5,
  ar_deviation = c(0.1, 0.49),
  ar_dist_type = c("uniform", "reciprocal")
)
}
\arguments{
\item{n}{number of instances to generate per class}

\item{no_of_points}{number of demand points in an instance}

\item{ar_mean}{the arrival rate mean (arrivals per minute)}

\item{ar_deviation}{the deviation from arrival rate mean}

\item{ar_dist_type}{Argument passed to \code{generate_2d_instance}}
}
\value{
nothing
}
\description{
Generate instances for experiment
}
