% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-experiment.R
\name{generate_simulations}
\alias{generate_simulations}
\title{Generate simulations based on solutions}
\usage{
generate_simulations(
  flight = c("zoned", "free"),
  free_max_dist = c("0", ".2", ".5", "no constraint")
)
}
\arguments{
\item{flight}{character vector of flight methods, currently \code{c("zoned","free")}.}

\item{free_max_dist}{max distances for use with free flight, currently \code{c("0", ".2", ".5", "no constraint")}.}
}
\value{
nothing
}
\description{
A simulation is generated and saved in the solution folder for specified flight method, and max_dist for free flight.
}
