% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-experiment.R
\name{generate_solutions}
\alias{generate_solutions}
\title{Generate solutions to instances}
\usage{
generate_solutions(
  methods = c("km", "wkm-flexclust", "wkm-swkm", "ga-safe", "ga-tot", "ga-otv"),
  no_of_centers = c(5, 15)
)
}
\arguments{
\item{methods}{character vector of methods to use for solution (currently: km, wkm)}

\item{no_of_centers}{integer vector of number of centers, or uavs, to include}
}
\value{
nothing
}
\description{
A solution is generated and saved in the solution folder for all combinations of methods and number of centers.
}
