% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_centroids.R
\name{grid_centroids}
\alias{grid_centroids}
\title{Generate grid centroids for a problem instance}
\usage{
grid_centroids(instance, dimension = 3)
}
\arguments{
\item{instance}{A list returned from generate_2d_instance}

\item{dimension}{The dimension of the grid}
}
\value{
A list
}
\description{
Generate grid centroids for a problem instance
}
\examples{
# WIP
}
