% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-experiment.R
\name{hexadec}
\alias{hexadec}
\title{hexadecimal of given size}
\usage{
hexadec(size = 64)
}
\arguments{
\item{size}{Length of hexadecimal}
}
\value{
A string
}
\description{
hexadecimal of given size
}
\examples{
hexadec()
}
