% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectives.R
\name{OTV}
\alias{OTV}
\title{Operation Time Variance}
\usage{
OTV(solution, tau = 1)
}
\arguments{
\item{solution}{A list generated by one of the solution functions}

\item{tau}{The service time per delivery}
}
\value{
A numeric value
}
\description{
Operation Time Variance
}
