% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_bases}
\alias{plot_bases}
\title{Plot base locations for solution}
\usage{
plot_bases(solution)
}
\arguments{
\item{solution}{A list returned from a solution function}
}
\value{
A ggplot object
}
\description{
Plot base locations for solution
}
\examples{
instance <- generate_2d_instance()
solution <- solve_kmeans(instance = instance, no_of_centers = 5)
plot_bases(solution = solution)
}
