% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_centroid}
\alias{plot_centroid}
\title{Plot centroid locations with demand points}
\usage{
plot_centroid(instance, centroids)
}
\arguments{
\item{instance}{A list returned from generate_2d_instance}

\item{centroids}{A list returned from grid_centroids}
}
\value{
A ggplot object
}
\description{
Plot centroid locations with demand points
}
\examples{
instance <- generate_2d_instance()
centroids <- grid_centroids(instance, dimension = 3)
plot_centroid(instance = instance, centroids = centroids)
}
