% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_point}
\alias{plot_point}
\title{Plot demand points for problem instance}
\usage{
plot_point(instance)
}
\arguments{
\item{instance}{A list returned from generate_2d_instance}
}
\value{
A ggplot object
}
\description{
Plot demand points for problem instance
}
\examples{
plot_point(instance = generate_2d_instance())
}
