% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectives.R
\name{SAFE}
\alias{SAFE}
\title{Smallest distance between demand points not in same zone}
\usage{
SAFE(solution)
}
\arguments{
\item{solution}{A solution object from a solve_* function}
}
\value{
A single number that is the smallest distance
}
\description{
Smallest distance between demand points not in same zone
}
