% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{safe_dist}
\alias{safe_dist}
\title{Calculate smallest distance between points not in same zone}
\usage{
safe_dist(A)
}
\arguments{
\item{A}{data.matrix produced from solve_ga function with obj == "SAFE"}
}
\description{
Calculate smallest distance between points not in same zone
}
