% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_dist}
\alias{sim_dist}
\title{Calculate euclidean distance for the safety distances}
\usage{
sim_dist(A)
}
\arguments{
\item{A}{A data.matrix produced in the simulation}
}
\description{
Calculate euclidean distance for the safety distances
}
