% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulation}
\alias{simulation}
\title{Simulation for a zoning solution}
\usage{
simulation(
  solution,
  seed = 1,
  n_replications = 1,
  flight = c("zoned", "free"),
  max_dist = 1e+06,
  LOS = 14400,
  warmup = 0,
  speed_agent = 0.25,
  verbose = F
)
}
\arguments{
\item{solution}{A list returned from one of the solution approaches}

\item{seed}{Passed to the set.seed function for reproducibility of results}

\item{n_replications}{Number of times to repeat the simulation}

\item{flight}{Whether to use hard zoning or soft zoning (free-flight)}

\item{max_dist}{The maximum distance a demand point can be of a base to be serviced by that base}

\item{LOS}{Length of simulation in seconds}

\item{warmup}{Fraction of LOS to discard as warmup period}

\item{speed_agent}{Movement per time unit for an agent}

\item{verbose}{Whether or not to print the information on the simulation}
}
\value{
None at the moment
}
\description{
Simulation for a zoning solution
}
\examples{
# WIP
}
