% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_ga.R
\name{solve_ga}
\alias{solve_ga}
\title{Genetic algorithm solution approach}
\usage{
solve_ga(
  instance,
  centroids,
  no_of_centers,
  obj = c("TOT", "SAFE"),
  miter = 10
)
}
\arguments{
\item{instance}{A list returned from generate_2d_instance}

\item{centroids}{A list returned from grid_centroids}

\item{no_of_centers}{The number of base locations to generate}

\item{obj}{Objective function to use (currently "TOT" or "SAFE")}

\item{miter}{Max number of iterations}
}
\value{
A list
}
\description{
Genetic algorithm solution approach
}
\examples{
# WIP
}
