% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_kmeans.R
\name{solve_kmeans}
\alias{solve_kmeans}
\title{KMeans solution approach}
\usage{
solve_kmeans(instance, no_of_centers = 5)
}
\arguments{
\item{instance}{A list returned from generate_2d_instance}

\item{no_of_centers}{The number of base locations to generate}
}
\value{
A list
}
\description{
KMeans solution approach
}
\examples{
instance <- generate_2d_instance()
solution <- solve_kmeans(instance = instance, no_of_centers = 5)
}
