% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectives.R
\name{TOT}
\alias{TOT}
\title{Total Operation Time}
\usage{
TOT(solution, tau = 0)
}
\arguments{
\item{solution}{A list generated by one of the solution functions}

\item{tau}{The service time per delivery}
}
\value{
A numeric value
}
\description{
Total Operation Time
}
