% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectives.R
\name{WCSS}
\alias{WCSS}
\title{Within-Cluster sum of squares}
\usage{
WCSS(solution)
}
\arguments{
\item{solution}{A list generated by one of the solution functions}
}
\value{
A numeric value
}
\description{
Within-Cluster sum of squares
}
