% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-experiment.R
\name{generate_instances}
\alias{generate_instances}
\title{Generate instances for experiment}
\usage{
generate_instances(
  n = 5,
  no_of_points = 100,
  ar_mean = 40,
  ar_deviation = c(10, 35)
)
}
\arguments{
\item{n}{number of instances to generate per class}

\item{no_of_points}{number of demand points in an instance}

\item{ar_mean}{the arrival rate mean}

\item{ar_deviation}{the deviation from arrival rate mean}
}
\value{
nothing
}
\description{
Generate instances for experiment
}
