% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_bases}
\alias{plot_bases}
\title{Plot base locations for KMeans solution}
\usage{
plot_bases(instance, solution)
}
\arguments{
\item{instance}{A list returned from generate_2d_instance}

\item{solution}{A list returned from solve_kmeans}
}
\value{
A ggplot object
}
\description{
Plot base locations for KMeans solution
}
\examples{
instance <- generate_2d_instance()
solution <- solve_kmeans(instance = instance, no_of_centers = 5)
plot_bases(instance = instance, solution = solution)
}
