% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulation}
\alias{simulation}
\title{Simulation for a zoning solution}
\usage{
simulation(
  solution,
  flight = c("zoned", "free"),
  log = F,
  max_dist = 1e+06,
  LOS = 600,
  warmup = 0
)
}
\arguments{
\item{solution}{A list returned from one of the solution approaches}

\item{flight}{Whether to use hard zoning or soft zoning (free-flight)}

\item{log}{Whether or not to log results}

\item{max_dist}{The maximum distance a demand point can be of a base to be serviced by that base}

\item{LOS}{Length of simulation in seconds}

\item{warmup}{fraction of LOS to discard as warmup period}
}
\value{
None at the moment
}
\description{
Simulation for a zoning solution
}
\examples{
# WIP
}
